examples/distributed/social-media/user-profile-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6

# OpenTelemetry packages
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-propagator-b3==1.21.0

# StatsD client
statsd==4.0.1

# Database migration
alembic==1.12.1

# JSON logging
python-json-logger==2.0.7
